changed HCI/HID interface implementation
- added keycode  modifier (alt, shift etc)
- refactored keypad- feature
- reworked keypad and keycode eventing
- removed keycode implementation
- added/refactored GPIO implementation
- refactored command serialization
- refactored parameter serialization
- fixed commands: hash, reset, temperature, display, feature, keypad, identity
- fixed varikey engine initialization
- fixed varikey app shutdown
- code hash tag size as variable

+++ src/engine/interface/hci/commander/control/cmd_hash_msg.hpp
@@ -39,7 +39,7 @@ namespace engine
                 struct message_t
                 {
                     RESULT result;
-                    uint32_t hash;
+                    uint16_t hash;
                     chunk_t value;
                 };
